Working scene-viewer on android
