+++ examples/scene-viewer/Cargo.toml
@@ -16,6 +16,9 @@ exclude = [
     "screenshot.jpg",
 ]
 
+[lib]
+crate-type = ["lib", "cdylib"]
+
 [[bin]]
 name = "scene-viewer"
 path = "src/main.rs"
@@ -45,9 +48,23 @@ tracy-client = { version = "0.12", optional = true }
 wgpu-profiler = "0.7.0"
 winit = "0.26"
 
-
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 tokio = { version = "1", features = ["rt", "rt-multi-thread"] }
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
 wasm-bindgen-futures = "0.4"
+
+[target.'cfg(target_os = "android")'.dependencies]
+ndk-glue = { version = "0.5", features = ["logger"] }
+
+[package.metadata.android]
+build_targets = ["aarch64-linux-android"]
+assets = "resources"
+
+[package.metadata.android.sdk]
+min_sdk_version = 28
+target_sdk_version = 28
+compile_sdk_version = 28
+
+[package.metadata.android.application.activity]
+orientation = "landscape"
